Update test configuration and refactor test cases for OpenAI chat integration
