+++ pyproject.toml
@@ -16,7 +16,7 @@ test = [
 package = true
 
 [tool.pytest.ini_options]
-testpaths = ["tests"]
+testpaths = ["src"]
 python_files = "test_*.py"
 python_classes = "Test*"
 python_functions = "test_*"
